+++ sigil.go
@@ -20,16 +20,16 @@ var (
 	rightDelim      string
 )
 
+var leftDelim = "{{"
+var rightDelim = "}}"
 var fnMap = template.FuncMap{}
 
 func init() {
-	leftDelim = os.Getenv("SIGIL_LEFT_DELIM")
-	if leftDelim == "" {
-		leftDelim = "{{"
-	}
-	rightDelim = os.Getenv("SIGIL_RIGHT_DELIM")
-	if rightDelim == "" {
-		rightDelim = "}}"
+	delims := os.Getenv("SIGIL_DELIMS")
+	if delims != "" {
+		d := strings.Split(delims, ",")
+		leftDelim = d[0]
+		rightDelim = d[1]
 	}
 }